+++ src/utils/tests/mod.rs
@@ -1,6 +1,7 @@
 mod data_decoded_utils;
 mod errors;
 mod json;
+mod macros;
 mod method_names;
 mod transactions;
 mod urls;
